+++ src/frontend/src/app/shared/models/declarative-form-modal-config.type.ts
@@ -1,11 +1,13 @@
 import { DeclarativeFormConfig } from '~/app/shared/models/declarative-form-config.type';
 
 export type DeclarativeFormModalConfig = {
-  title: string;
+  title?: string;
+  subtitleIconClass?: string;
   subtitle?: string;
   submitButtonVisible?: boolean; // Defaults to `true`
   submitButtonText?: string; // Defaults to `OK`
   submitButtonResult?: any; // Defaults to form values
+  submitButtonClass?: string;
   cancelButtonVisible?: boolean; // Defaults to `true`
   cancelButtonText?: string; // Defaults to `Cancel`
   cancelButtonResult?: any; // Defaults to `false`